slight changes to seg2gold method

+++ transforms/cpp-seg2gold.h
@@ -1,1 +1,1 @@
-long *CppMapping(long *segmentation, int *gold, long nentries, double match_threshold, double nonzero_threshold);
+long *CppMapping(long *segmentation, long *gold, long nentries, double match_threshold, double nonzero_threshold);
